The rgb() color game
A browser based game I built in conjunction with a course on Udemy. The
game will display an rgb() color value in text and asks the user to
pick the right color from a series of squares displaying different
colors. The number of squares varies - either 3 or 6 depending on the
difficultly as selected by the user.